Remove `include/grpcpp/impl/codegen/client_callback.h` (#31005)
