+++ src/compiler/cpp_generator.cc
@@ -140,7 +140,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/generic/async_generic_service.h",
         "grpcpp/support/async_stream.h",
         "grpcpp/support/async_unary_call.h",
-        "grpcpp/impl/codegen/client_callback.h",
+        "grpcpp/support/client_callback.h",
         "grpcpp/impl/codegen/client_context.h",
         "grpcpp/impl/codegen/completion_queue.h",
         "grpcpp/impl/codegen/message_allocator.h",
@@ -1652,7 +1652,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "grpcpp/support/async_unary_call.h",
         "grpcpp/impl/codegen/channel_interface.h",
         "grpcpp/impl/codegen/client_unary_call.h",
-        "grpcpp/impl/codegen/client_callback.h",
+        "grpcpp/support/client_callback.h",
         "grpcpp/impl/codegen/message_allocator.h",
         "grpcpp/impl/codegen/method_handler.h",
         "grpcpp/impl/codegen/rpc_service_method.h",
