+++ test/cpp/codegen/compiler_test_golden
@@ -30,7 +30,7 @@
 #include <grpcpp/generic/async_generic_service.h>
 #include <grpcpp/support/async_stream.h>
 #include <grpcpp/support/async_unary_call.h>
-#include <grpcpp/impl/codegen/client_callback.h>
+#include <grpcpp/support/client_callback.h>
 #include <grpcpp/impl/codegen/client_context.h>
 #include <grpcpp/impl/codegen/completion_queue.h>
 #include <grpcpp/impl/codegen/message_allocator.h>
